requirements.txt
alembic==1.8.1
async-timeout==4.0.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
urllib3==1.26.12
uuid==1.30
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
